third_party/WebKit/LayoutTests/performance/event-timing/event-timing-crossiframe.html

<!DOCTYPE html>
<html>

<head>
  <meta charset=utf-8 />
  <title>Event Timing: entries should be observable by its own frame.</title>
</head>

<body>
  <button id='button' onclick='1'>Generate a 'click' event</button>
</body>
<div>
  <iframe src=./resources/event-timing-crossiframe-childframe.html></iframe>
</div>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script src=./resources/event-timing-support.js></script>
<script>
  let clickTimeMin;
  let processingStartMin;

  function validateEntries() {
    const entries = performance.getEntriesByName('click');
    const onloadTime = performance.timing.loadEventStart - performance.timeOrigin;

    assert_equals(entries.length, 1,
      "Observer of main frames should only capture main-frame event-timing entries."
    );
    const entry = entries[0];
    verifyClickEvent(entry);

    assert_true(entry.processingStart > processingStartMin);
    assert_true(entry.processingStart < onloadTime);
    assert_true(entry.startTime > clickTimeMin);
    assert_true(entry.startTime < onloadTime);
  }

  function validateChildFrameEntries(msg) {
    const childFrameData = msg.data;
    assert_equals(childFrameData.bufferedEntries.length, 1,
      "Event Timing of child frames should only capture child-frame event-timing entries."
    );
    const entry = entries[0];
    verifyClickEvent(entry);

    assert_true(entry.processingStart > childFrameData.processingStartMin);
    assert_true(entry.processingStart < childFrameData.onloadTime);
    assert_true(entry.startTime > childFrameData.clickTimeMin);
    assert_true(entry.startTime < childFrameData.onloadTime);

    assert_array_equals(childFrameData.observedEntries, childFrameData.bufferedEntries);
  }

  async_test(function(t) {
    clickTimeMin = performance.now();
    clickAndBlockMain('button');
    processingStartMin = performance.now();
    const childFrameEntriesPromise = new Promise((resolve, reject) => {
      window.addEventListener("message", (event) => {
        resolve(event.data);
      }, false);
    });
    on_event(window, 'load', e => {
      childFrameEntriesPromise.then((entries) => {
        t.step(() => {
          validateChildFrameEntries(entries);
          validateEntries();
        });
        t.done();
      });
    });
  }, "Event Timing: entries should only be observable by its own frame.");

</script>

</html>
